// LineupOptimizer/src/app/shared/models/Team.ts
export class Team {
    _id: string = '';
    name: string = '';
    gamesPlayed: number = 0.0;
    points: number = 0.0;
    rebounds: number = 0.0;
    assists: number = 0.0;
    steals: number = 0.0;
    blocks: number = 0.0;
    turnovers: number = 0.0;
    pf: number = 0.0;
    defRbds: number = 0.0;
    offRbds: number = 0.0;
    totRbds: number = 0.0;
    fgMade: number = 0.0;
    fgatt: number = 0.0;
    fgPct: number = 0.0;
    threesMade: number = 0.0;
    threesAtt: number = 0.0;
    threePct: number = 0.0;
    twosMade: number = 0.0;
    twosAtt: number = 0.0;
    twoPct: number = 0.0;
    ftMade: number = 0.0;
    ftAtt: number = 0.0;
    ftPct: number = 0.0;
}